+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+-   Websocket based connection
+
+### Breaking Changes
+
+-   The `Error` type has been split up into more specific error types dependent on the connection used.
+-   Client is now generic over the error type of query/mutate/subscribe
+-   Auth related methods now use `AuthError`
+-   Http connection method now use `HttpConnectionError`
+
 ## v0.2.0 (2021-03-23)
 
 ### Added